requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
rich>=12.0.0
customtkinter>=5.0.0
Pillow>=9.0.0
pycryptodome>=3.15.0
requests>=2.25.0